Moved CumulativeTexture into the renderable folder
